Add `cf` ("control-flow") module and split `cfg` into `cf::unstructured` vs `cf::structurize`.
