+++ src/spv/lift.rs
@@ -1,5 +1,6 @@
 //! SPIR-T to SPIR-V lifting.
 
+use crate::cf::{self, SelectionKind};
 use crate::func_at::FuncAt;
 use crate::spv::{self, spec};
 use crate::visit::{InnerVisit, Visitor};
@@ -7,8 +8,8 @@ use crate::{
     AddrSpace, Attr, AttrSet, Const, ConstDef, ConstKind, Context, DataInst, DataInstDef,
     DataInstKind, DbgSrcLoc, DeclDef, EntityList, ExportKey, Exportee, Func, FuncDecl, FuncParam,
     FxIndexMap, FxIndexSet, GlobalVar, GlobalVarDefBody, Import, Module, ModuleDebugInfo,
-    ModuleDialect, Node, NodeKind, NodeOutputDecl, OrdAssertEq, Region, RegionInputDecl,
-    SelectionKind, Type, TypeDef, TypeKind, TypeOrConst, Value, cfg,
+    ModuleDialect, Node, NodeKind, NodeOutputDecl, OrdAssertEq, Region, RegionInputDecl, Type,
+    TypeDef, TypeKind, TypeOrConst, Value,
 };
 use rustc_hash::FxHashMap;
 use smallvec::SmallVec;
@@ -262,7 +263,7 @@ struct FuncLifting<'a> {
 /// What determines the values for [`Value::RegionInput`]s, for a specific
 /// region (effectively the subset of "region parents" that support inputs).
 ///
-/// Note that this is not used when a [`cfg::ControlInst`] has `target_inputs`,
+/// Note that this is not used when a [`cf::unstructured::ControlInst`] has `target_inputs`,
 /// and the target [`Region`] itself has phis for its `inputs`.
 enum RegionInputsSource {
     FuncParams,
@@ -299,7 +300,7 @@ struct Phi {
     default_value: Option<Value>,
 }
 
-/// Similar to [`cfg::ControlInst`], except:
+/// Similar to [`cf::unstructured::ControlInst`], except:
 /// * `targets` use [`CfgPoint`]s instead of [`Region`]s, to be able to
 ///   reach any of the SPIR-V blocks being created during lifting
 /// * φ ("phi") values can be provided for targets regardless of "which side" of
@@ -310,7 +311,7 @@ struct Phi {
 struct Terminator<'a> {
     attrs: AttrSet,
 
-    kind: Cow<'a, cfg::ControlInstKind>,
+    kind: Cow<'a, cf::unstructured::ControlInstKind>,
 
     // FIXME(eddyb) use `Cow` or something, but ideally the "owned" case always
     // has at most one input, so allocating a whole `Vec` for that seems unwise.
@@ -640,8 +641,13 @@ impl<'a> FuncLifting<'a> {
                         .as_ref()
                         .and_then(|cfg| cfg.control_inst_on_exit_from.get(region));
                     if let Some(terminator) = unstructured_terminator {
-                        let cfg::ControlInst { attrs, kind, inputs, targets, target_inputs } =
-                            terminator;
+                        let cf::unstructured::ControlInst {
+                            attrs,
+                            kind,
+                            inputs,
+                            targets,
+                            target_inputs,
+                        } = terminator;
                         Terminator {
                             attrs: *attrs,
                             kind: Cow::Borrowed(kind),
@@ -664,7 +670,7 @@ impl<'a> FuncLifting<'a> {
                         assert!(region == func_def_body.body);
                         Terminator {
                             attrs: AttrSet::default(),
-                            kind: Cow::Owned(cfg::ControlInstKind::Return),
+                            kind: Cow::Owned(cf::unstructured::ControlInstKind::Return),
                             inputs: func_def_body.at_body().def().outputs.clone(),
                             targets: [].into_iter().collect(),
                             target_phi_values: FxIndexMap::default(),
@@ -683,7 +689,9 @@ impl<'a> FuncLifting<'a> {
 
                         NodeKind::Select { kind, scrutinee, cases } => Terminator {
                             attrs: AttrSet::default(),
-                            kind: Cow::Owned(cfg::ControlInstKind::SelectBranch(kind.clone())),
+                            kind: Cow::Owned(cf::unstructured::ControlInstKind::SelectBranch(
+                                kind.clone(),
+                            )),
                             inputs: [*scrutinee].into_iter().collect(),
                             targets: cases
                                 .iter()
@@ -696,7 +704,7 @@ impl<'a> FuncLifting<'a> {
                         NodeKind::Loop { initial_inputs: _, body, repeat_condition: _ } => {
                             Terminator {
                                 attrs: AttrSet::default(),
-                                kind: Cow::Owned(cfg::ControlInstKind::Branch),
+                                kind: Cow::Owned(cf::unstructured::ControlInstKind::Branch),
                                 inputs: [].into_iter().collect(),
                                 targets: [CfgPoint::RegionEntry(*body)].into_iter().collect(),
                                 target_phi_values: FxIndexMap::default(),
@@ -716,7 +724,9 @@ impl<'a> FuncLifting<'a> {
 
                         NodeKind::ExitInvocation { kind, inputs } => Terminator {
                             attrs: AttrSet::default(),
-                            kind: Cow::Owned(cfg::ControlInstKind::ExitInvocation(kind.clone())),
+                            kind: Cow::Owned(cf::unstructured::ControlInstKind::ExitInvocation(
+                                kind.clone(),
+                            )),
                             inputs: inputs.clone(),
                             targets: [].into_iter().collect(),
                             target_phi_values: FxIndexMap::default(),
@@ -743,7 +753,7 @@ impl<'a> FuncLifting<'a> {
 
                         NodeKind::Select { .. } => Terminator {
                             attrs: AttrSet::default(),
-                            kind: Cow::Owned(cfg::ControlInstKind::Branch),
+                            kind: Cow::Owned(cf::unstructured::ControlInstKind::Branch),
                             inputs: [].into_iter().collect(),
                             targets: [parent_exit].into_iter().collect(),
                             target_phi_values: region_outputs
@@ -775,7 +785,7 @@ impl<'a> FuncLifting<'a> {
                             if is_infinite_loop {
                                 Terminator {
                                     attrs: AttrSet::default(),
-                                    kind: Cow::Owned(cfg::ControlInstKind::Branch),
+                                    kind: Cow::Owned(cf::unstructured::ControlInstKind::Branch),
                                     inputs: [].into_iter().collect(),
                                     targets: [backedge].into_iter().collect(),
                                     target_phi_values,
@@ -784,9 +794,11 @@ impl<'a> FuncLifting<'a> {
                             } else {
                                 Terminator {
                                     attrs: AttrSet::default(),
-                                    kind: Cow::Owned(cfg::ControlInstKind::SelectBranch(
-                                        SelectionKind::BoolCond,
-                                    )),
+                                    kind: Cow::Owned(
+                                        cf::unstructured::ControlInstKind::SelectBranch(
+                                            SelectionKind::BoolCond,
+                                        ),
+                                    ),
                                     inputs: [repeat_condition].into_iter().collect(),
                                     targets: [backedge, parent_exit].into_iter().collect(),
                                     target_phi_values,
@@ -801,7 +813,7 @@ impl<'a> FuncLifting<'a> {
                 // implied edge from a `Block`'s `Entry` to its `Exit`).
                 (_, Some(succ_cursor)) => Terminator {
                     attrs: AttrSet::default(),
-                    kind: Cow::Owned(cfg::ControlInstKind::Branch),
+                    kind: Cow::Owned(cf::unstructured::ControlInstKind::Branch),
                     inputs: [].into_iter().collect(),
                     targets: [succ_cursor.point].into_iter().collect(),
                     target_phi_values: FxIndexMap::default(),
@@ -874,8 +886,10 @@ impl<'a> FuncLifting<'a> {
             // SPIR-V allows their targets to just be the whole merge block
             // (the same one that `OpSelectionMerge` describes).
             let block = &blocks[block_idx];
-            if let (cfg::ControlInstKind::SelectBranch(_), Some(Merge::Selection(merge_point))) =
-                (&*block.terminator.kind, block.terminator.merge)
+            if let (
+                cf::unstructured::ControlInstKind::SelectBranch(_),
+                Some(Merge::Selection(merge_point)),
+            ) = (&*block.terminator.kind, block.terminator.merge)
             {
                 for target_idx in 0..block.terminator.targets.len() {
                     let block = &blocks[block_idx];
@@ -902,7 +916,7 @@ impl<'a> FuncLifting<'a> {
                         (phis.is_empty()
                             && insts.iter().all(|insts| insts.is_empty())
                             && *attrs == AttrSet::default()
-                            && matches!(**kind, cfg::ControlInstKind::Branch)
+                            && matches!(**kind, cf::unstructured::ControlInstKind::Branch)
                             && inputs.is_empty()
                             && targets.len() == 1
                             && target_phi_values.is_empty()
@@ -925,7 +939,7 @@ impl<'a> FuncLifting<'a> {
                     &block.terminator;
 
                 (*attrs == AttrSet::default()
-                    && matches!(**kind, cfg::ControlInstKind::Branch)
+                    && matches!(**kind, cf::unstructured::ControlInstKind::Branch)
                     && inputs.is_empty()
                     && targets.len() == 1
                     && target_phi_values.is_empty()
@@ -951,7 +965,7 @@ impl<'a> FuncLifting<'a> {
                             new_terminator,
                             Terminator {
                                 attrs: Default::default(),
-                                kind: Cow::Owned(cfg::ControlInstKind::Unreachable),
+                                kind: Cow::Owned(cf::unstructured::ControlInstKind::Unreachable),
                                 inputs: Default::default(),
                                 targets: Default::default(),
                                 target_phi_values: Default::default(),
@@ -1330,25 +1344,25 @@ impl LazyInst<'_, '_> {
             },
             Self::Terminator { parent_func, terminator } => {
                 let inst = match &*terminator.kind {
-                    cfg::ControlInstKind::Unreachable => wk.OpUnreachable.into(),
-                    cfg::ControlInstKind::Return => {
+                    cf::unstructured::ControlInstKind::Unreachable => wk.OpUnreachable.into(),
+                    cf::unstructured::ControlInstKind::Return => {
                         if terminator.inputs.is_empty() {
                             wk.OpReturn.into()
                         } else {
                             wk.OpReturnValue.into()
                         }
                     }
-                    cfg::ControlInstKind::ExitInvocation(cfg::ExitInvocationKind::SpvInst(
+                    cf::unstructured::ControlInstKind::ExitInvocation(
+                        cf::ExitInvocationKind::SpvInst(inst),
+                    )
+                    | cf::unstructured::ControlInstKind::SelectBranch(SelectionKind::SpvInst(
                         inst,
                     )) => inst.clone(),
 
-                    cfg::ControlInstKind::Branch => wk.OpBranch.into(),
+                    cf::unstructured::ControlInstKind::Branch => wk.OpBranch.into(),
 
-                    cfg::ControlInstKind::SelectBranch(SelectionKind::BoolCond) => {
+                    cf::unstructured::ControlInstKind::SelectBranch(SelectionKind::BoolCond) => {
                         wk.OpBranchConditional.into()
-                    }
-                    cfg::ControlInstKind::SelectBranch(SelectionKind::SpvInst(inst)) => {
-                        inst.clone()
                     }
                 };
                 spv::InstWithIds {
